// Frontend/SchoolManagementSystem/src/app/models/principal/course-list-for-class-subject.ts
export interface ICourseListForClassSubjectInterface {
    courseId: number;
    chapterSerialNumber: number;
    chapterName: string;
    probableWeightageInExam: number;
    probableDurationToTeach: string;
    isOptionalToTeach: boolean;
    learningObjectives: string;
    classId: number;
    subjectId: number;
    className: string;
    subjectName: string;
  }
  